全商品用のコントローラーとPC用とBook用のコントローラーに切り分けた
