+++ backend/src/main/java/com/example/controller/ProductController.java
@@ -1,6 +1,5 @@
 package com.example.controller;
 
-import com.example.dto.request.AddPCRequest;
 import com.example.dto.response.BookDetailResponse;
 import com.example.dto.response.PCDetailResponse;
 import com.example.dto.response.ProductsResponse;
@@ -11,13 +10,10 @@ import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-/** 商品操作を行うコントローラクラス. */
+/** 全商品の操作を行うコントローラクラス. */
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/products")
@@ -97,94 +93,5 @@ public class ProductController {
     List<ProductsResponse> pagedList = productsList.subList(fromIndex, toIndex);
 
     return ResponseEntity.ok(pagedList);
-  }
-
-  /**
-   * PC一覧を取得するエンドポイント
-   *
-   * @param sort ソート条件
-   * @param limit 1ページあたりの表示件数
-   * @param offset 次ページの開始位置
-   * @param keyword 検索キーワード
-   * @return PC一覧結果
-   */
-  @GetMapping("/pcs")
-  public ResponseEntity<?> getPCs(
-      @RequestParam(defaultValue = "priceAsc") String sort,
-      @RequestParam(defaultValue = "0") Integer limit,
-      @RequestParam(defaultValue = "20") Integer offset,
-      @RequestParam(defaultValue = "") String keyword) {
-    Sort sorting =
-        switch (sort) {
-          case "priceAsc" -> Sort.by(Sort.Direction.ASC, "price");
-          case "priceDesc" -> Sort.by(Sort.Direction.DESC, "price");
-          default -> Sort.by("id");
-        };
-    Pageable pageable = PageRequest.of(limit, offset, sorting);
-
-    return ResponseEntity.ok(pcService.findPcs(keyword, pageable));
-  }
-
-  /**
-   * PCをPCsテーブルに追加するエンドポイント.
-   *
-   * @param request PC登録リクエスト
-   * @return 登録されたPC情報
-   */
-  @PostMapping("/pcs")
-  public ResponseEntity<?> addPCToTable(@RequestBody AddPCRequest request) {
-    Pc pc = new Pc();
-    pc.setName(request.getName());
-    pc.setPrice(request.getPrice());
-    pc.setMemory(request.getMemory());
-    pc.setStorage(request.getStorage());
-    pc.setDeviceSize(request.getDeviceSize());
-    pc.setDeviceType(request.getDeviceType());
-
-    Os os = new Os();
-    os.setId(request.getOsId());
-    pc.setOs(os);
-
-    Cpu cpu = new Cpu();
-    cpu.setId(request.getCpuId());
-    pc.setCpu(cpu);
-
-    Gpu gpu = new Gpu();
-    gpu.setId(request.getGpuId());
-    pc.setGpu(gpu);
-
-    Purpose purpose = new Purpose();
-    purpose.setId(request.getPurposeId());
-    pc.setPurpose(purpose);
-
-    pcService.registerPC(pc);
-
-    return ResponseEntity.ok().build();
-  }
-
-  /**
-   * Book一覧を取得するエンドポイント
-   *
-   * @param sort ソート条件
-   * @param limit 1ページあたりの表示件数
-   * @param offset 次ページの開始位置
-   * @param keyword 検索キーワード
-   * @return Book一覧結果
-   */
-  @GetMapping("/books")
-  public ResponseEntity<?> getBooks(
-      @RequestParam(defaultValue = "priceAsc") String sort,
-      @RequestParam(defaultValue = "0") Integer limit,
-      @RequestParam(defaultValue = "20") Integer offset,
-      @RequestParam(defaultValue = "") String keyword) {
-    Sort sorting =
-        switch (sort) {
-          case "priceAsc" -> Sort.by(Sort.Direction.ASC, "price");
-          case "priceDesc" -> Sort.by(Sort.Direction.DESC, "price");
-          default -> Sort.by("id");
-        };
-    Pageable pageable = PageRequest.of(limit, offset, sorting);
-
-    return ResponseEntity.ok(bookService.findBooks(keyword, pageable));
   }
 }
